backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
tinydb==4.8.0
fastapi-mail==1.4.1
cryptography==42.0.0
pydantic[email]==2.5.3
